+++ src/zephyr/client.ts
@@ -9,6 +9,7 @@ import { ApiClient } from "./common/api-client";
 import { GetEnvironments } from "./tool/environment/get-environments";
 import { CreateFolder } from "./tool/folder/create-folder";
 import { GetTestCases as GetIssueLinkTestCases } from "./tool/issue-link/get-test-cases";
+import { GetTestCycles as GetIssueLinkTestCycles } from "./tool/issue-link/get-test-cycles";
 import { GetPriorities } from "./tool/priority/get-priorities";
 import { GetProject } from "./tool/project/get-project";
 import { GetProjects } from "./tool/project/get-projects";
@@ -111,6 +112,7 @@ export class ZephyrClient implements Client {
       new CreateTestExecutionIssueLink(this),
       new GetTestCaseSteps(this),
       new GetIssueLinkTestCases(this),
+      new GetIssueLinkTestCycles(this),
       new GetTestScript(this),
       new CreateTestCycleWebLink(this),
       new GetTestExecutionSteps(this),